Fix issue with update record not updating inverse relationships properly
