+++ packages/@orbit/record-cache/test/sync-record-cache-test.ts
@@ -158,30 +158,60 @@ module('SyncRecordCache', function(hooks) {
     );
   });
 
-  test('#patch updates inverse hasOne relationship when a record with relationships unspecified is added', function(assert) {
+  test('#patch updates inverse hasOne relationship when a record with relationships unspecified is added - record added after', function(assert) {
     const cache = new Cache({ schema, keyMap });
 
     const jupiter: Record = { type: 'planet', id: 'p1', attributes: { name: 'Jupiter' }, relationships: { moons: { data: [{ type: 'moon', id: 'm1' }] } } };
     const io: Record = { type: 'moon', id: 'm1', attributes: { name: 'Io' }};
 
     cache.patch(t => [
-      t.addRecord(jupiter),
-      t.addRecord(io)
+      t.updateRecord(jupiter),
+      t.updateRecord(io)
     ]);
 
     assert.deepEqual(cache.getRecordSync({ type: 'planet', id: 'p1' }).relationships.moons.data, [{ type: 'moon', id: 'm1' }], 'Jupiter has been assigned to Io');
     assert.deepEqual(cache.getRecordSync({ type: 'moon', id: 'm1' }).relationships.planet.data, { type: 'planet', id: 'p1' }, 'Io has been assigned to Jupiter');
   });
 
-  test('#patch updates inverse hasMany relationship when a record with relationships unspecified is added', function(assert) {
+  test('#patch updates inverse hasOne relationship when a record with relationships unspecified is added - record added before', function(assert) {
+    const cache = new Cache({ schema, keyMap });
+
+    const jupiter: Record = { type: 'planet', id: 'p1', attributes: { name: 'Jupiter' }, relationships: { moons: { data: [{ type: 'moon', id: 'm1' }] } } };
+    const io: Record = { type: 'moon', id: 'm1', attributes: { name: 'Io' }};
+
+    cache.patch(t => [
+      t.updateRecord(io),
+      t.updateRecord(jupiter)
+    ]);
+
+    assert.deepEqual(cache.getRecordSync({ type: 'planet', id: 'p1' }).relationships.moons.data, [{ type: 'moon', id: 'm1' }], 'Jupiter has been assigned to Io');
+    assert.deepEqual(cache.getRecordSync({ type: 'moon', id: 'm1' }).relationships.planet.data, { type: 'planet', id: 'p1' }, 'Io has been assigned to Jupiter');
+  });
+
+  test('#patch updates inverse hasMany relationship when a record with relationships unspecified is added - record added after', function(assert) {
     const cache = new Cache({ schema, keyMap });
 
     const io: Record = { type: 'moon', id: 'm1', attributes: { name: 'Io' }, relationships: { planet: { data: { type: 'planet', id: 'p1'} } } };
     const jupiter: Record = { type: 'planet', id: 'p1', attributes: { name: 'Jupiter' } };
 
     cache.patch(t => [
-      t.addRecord(io),
-      t.addRecord(jupiter)
+      t.updateRecord(io),
+      t.updateRecord(jupiter)
+    ]);
+
+    assert.deepEqual(cache.getRecordSync({ type: 'planet', id: 'p1' }).relationships.moons.data, [{ type: 'moon', id: 'm1' }], 'Jupiter has been assigned to Io');
+    assert.deepEqual(cache.getRecordSync({ type: 'moon', id: 'm1' }).relationships.planet.data, { type: 'planet', id: 'p1' }, 'Io has been assigned to Jupiter');
+  });
+
+  test('#patch updates inverse hasMany relationship when a record with relationships unspecified is added - record added before', function(assert) {
+    const cache = new Cache({ schema, keyMap });
+
+    const io: Record = { type: 'moon', id: 'm1', attributes: { name: 'Io' }, relationships: { planet: { data: { type: 'planet', id: 'p1'} } } };
+    const jupiter: Record = { type: 'planet', id: 'p1', attributes: { name: 'Jupiter' } };
+
+    cache.patch(t => [
+      t.updateRecord(jupiter),
+      t.updateRecord(io)
     ]);
 
     assert.deepEqual(cache.getRecordSync({ type: 'planet', id: 'p1' }).relationships.moons.data, [{ type: 'moon', id: 'm1' }], 'Jupiter has been assigned to Io');
@@ -195,8 +225,8 @@ module('SyncRecordCache', function(hooks) {
     const jupiter: Record = { type: 'planet', id: 'p1', attributes: { name: 'Jupiter' }, relationships: { moons: { data: [] }}};
 
     cache.patch(t => [
-      t.addRecord(io),
-      t.addRecord(jupiter)
+      t.updateRecord(io),
+      t.updateRecord(jupiter)
     ]);
 
     assert.deepEqual(cache.getRecordSync({ type: 'planet', id: 'p1' }).relationships.moons.data, [], 'Jupiter has been assigned to Io');
@@ -210,8 +240,8 @@ module('SyncRecordCache', function(hooks) {
     const io: Record = { type: 'moon', id: 'm1', attributes: { name: 'Io' }, relationships: { planet: { data: null } } };
 
     cache.patch(t => [
-      t.addRecord(jupiter),
-      t.addRecord(io)
+      t.updateRecord(jupiter),
+      t.updateRecord(io)
     ]);
 
     assert.deepEqual(cache.getRecordSync({ type: 'planet', id: 'p1' }).relationships.moons.data, [], 'Jupiter has been assigned to Io');
